+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes after version 1.6.16 to this project will be documented in t
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
+### Fixed
+- eager instantiation of WordPress factories causing warnings (#281)
 
 ## [2.2.16] 2019-07-23;
 ### Added